Time: 0 ms (100%), Space: 44.1 MB (38.6%) - LeetHub

+++ 0789-escape-the-ghosts/0789-escape-the-ghosts.java
@@ -1,10 +1,10 @@
 class Solution {
     public boolean escapeGhosts(int[][] ghosts, int[] target) {
 
-        int myDist = (int)(Math.abs(target[0]) + Math.abs(target[1]));
+        int myDist = Math.abs(target[0]) + Math.abs(target[1]);
     
         for(int i = 0; i < ghosts.length; i++){
-            int ghostD = (int)(Math.abs(ghosts[i][0] - target[0]) + Math.abs(ghosts[i][1] - target[1]));
+            int ghostD = Math.abs(ghosts[i][0] - target[0]) + Math.abs(ghosts[i][1] - target[1]);
             if(ghostD <= myDist) return false;
         }
